LayoutTests/imported/w3c/web-platform-tests/websockets/Create-protocols-repeated.any-expected.txt

CONSOLE MESSAGE: line 61: WebSocket protocols contain duplicates: 'echo'

PASS Create WebSocket - Pass a valid URL and an array of protocol strings with repeated values - SYNTAX_ERR is thrown 
